requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
langchain
langchain-openai
requests
httpx
pydantic